src/requirements.txt
streamlit
loguru
pandas
numpy
git+https://github.com/openai/whisper.git
cattrs
attrs
audio-recorder-streamlit